ai-advisor/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
psycopg[binary]==3.2.10
openai==1.55.3
python-dotenv==1.0.0
